Move test metadata generation into util.py
